refactor: ID 生成の共有ユーティリティ化と各所からの直接参照
