+++ src/main/utils/crypto.ts
@@ -1,7 +1,5 @@
 import { createHash } from 'node:crypto';
 
-export { generateId } from '@shared/utils/id';
-
 /**
  * 指定されたバイナリデータのMD5ハッシュ値を計算します。
  * @param buffer 計算対象のバイナリデータ
