+++ src/shared/utils/id.test.ts
@@ -1,0 +1,19 @@
+import { describe, it, expect } from 'vitest';
+import { generateId } from './id';
+
+describe('ID ユーティリティ', () => {
+  describe('generateId', () => {
+    it('UUID 形式の文字列を生成すること', () => {
+      const id = generateId();
+      // UUID v4 の形式チェック (8-4-4-4-12)
+      const uuidRegex = /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+      expect(id).toMatch(uuidRegex);
+    });
+
+    it('呼び出すたびに異なる ID を生成すること', () => {
+      const id1 = generateId();
+      const id2 = generateId();
+      expect(id1).not.toBe(id2);
+    });
+  });
+});
